exporters/pyproject.toml
[tool.poetry]
name = "planning-data-exporter"
version = "0.1.0"
description = "A hierarchical exporter for planning data"
authors = ["Albert J. Orlando <albertjorlando+github@gmail.com>"]
readme = "README.md"
packages = [
    {include = "todoist_data_exporter", from = "src"},
    {include = "google_calendar_exporter", from = "src"},
    {include = "planning_data_exporter", from = "src"}
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.0"
requests = "^2.31.0"
typing-extensions = "^4.10.0"
click = "^8.1.7"
rich = "^13.7.0"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.100.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
mypy = "^1.9.0"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
types-requests = "^2.32.0.20250328"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-v -s --cov-report term-missing"

[tool.poetry.scripts]
planning-export = "planning_data_exporter.cli:main"
todoist-export = "todoist_data_exporter.cli:main"
gcal-export = "google_calendar_exporter.main:run_export"
